third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/workers/shared-worker-secure-context.https-expected.txt

CONSOLE WARNING: line 27: Mixed Content: The page at 'https://127.0.0.1:8443/workers/shared-worker-secure-context.https.html' was loaded over HTTPS, but requested an insecure resource 'http://example.test:8000/workers/resources/load-worker-in-iframe.html'. This content should also be served over HTTPS.
CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/workers/shared-worker-secure-context.https.html' was loaded over HTTPS, but requested an insecure script 'http://example.test:8000/resources/get-host-info.js'. This content should also be served over HTTPS.
This is a testharness.js-based test.
PASS Shared workers secure context 
Harness: the test ran to completion.
